Teleprompter: Ensure that the prompters calculations
The prompter will double check that the width hasn't changed before starting prompting.

I also added a public `prepare()` method that signals the prompter that it should calculate widths again. This can be called as a signal to the prompter that you have made all viewport size adjustments and the environment will not change before the next prompting starts. This doesn't prevent a double check before we start, but it does ensure that a more costly calculation will not run when the first one sees changes.