RequestsService: use the this.accessToken value outside the promise...
I am not 100% sure, but I think this solves the [problem](https://github.com/flixpressllc/FrixxerNewsEditor/issues/33).

While debugging the issue, I added console.logs all over the place and discovered that the problem just vanished. As such, my guess is that however the `get` functions are transpiled leaves room for error when they are accessed inside a Promise constructor.

The reason I am thinking that is because they worked when I ran the project locally, but when it was built for production to the server, it failed (until the console.log statements were added).

In any case, the function is simpler and the logic is easier to follow anyway.